Only support USB driver for ESP32S targets
