+++ TactilityHeadless/Source/hal/usb/UsbTusb.h
@@ -1,0 +1,5 @@
+#pragma once
+
+bool tusbIsSupported();
+bool tusbStartMassStorageWithSdmmc();
+void tusbStop();
