+++ sdkconfig.board.yellow-board
@@ -24,8 +24,6 @@ CONFIG_PARTITION_TABLE_CUSTOM_FILENAME="partitions.csv"
 CONFIG_PARTITION_TABLE_FILENAME="partitions.csv"
 CONFIG_ELF_LOADER_CUSTOMER_SYMBOLS=y
 CONFIG_FATFS_LFN_HEAP=y
-CONFIG_TINYUSB_MSC_ENABLED=y
-CONFIG_TINYUSB_MSC_MOUNT_PATH="/sdcard"
 
 # Hardware: Main
 CONFIG_TT_BOARD_YELLOW_BOARD_24_CAP=y
